Update movie.js
Overall, building MoveList was a great way to practice my JavaScript, HTML, CSS, MySQL, and Bootstrap skills. By building a web application with dynamic functionality for searching movies names, as well as persisting the data in a database, I was able to gain a better understanding of these technologies and how they can be used together to create interactive web applications.